refactor(probability): share canonical exact values across APIs
